requirements.txt
flask
google-cloud-aiplatform
google-cloud-discoveryengine
langchain-google-vertexai
langchain-google-community
python-dotenv